@@ -1,85 +1,69 @@
 {
-    "name": "react-pro",
-    "version": "1.0.0",
-    "description": "react-pro-app",
-    "author": "koringz",
-    "license": "ISC",
-    "scripts": {
-        "open": "yarn go --open",
-        "go": "webpack-dev-server --config config/webpack.development.config.js",
-        "build:dev": "cross-env FIRST_ENV=development --env.NODE_ENV=development webpack --config config/webpack.development.config.js",
-        "build:prod": "cross-env FIRST_ENV=production --env.NODE_ENV=production webpack --config config/webpack.production.config.js",
-        "build:test": "cross-env FIRST_ENV=testing --env.NODE_ENV=testing webpack --config config/webpack.testing.config.js",
-        "plop:build": "plop pages `call name generator pages`",
-        "upgrade": "yarn upgrade-interactive --latest"
-    },
-    "dependencies": {
-        "@ant-design/icons": "^4.7.0",
-        "@babel/core": "^7.15.5",
-        "@babel/plugin-syntax-typescript": "^7.14.5",
-        "@babel/preset-env": "^7.15.4",
-        "@babel/preset-react": "^7.14.5",
-        "@babel/preset-typescript": "^7.15.0",
-        "@types/lodash": "^4.14.177",
-        "@types/react": "^17.0.20",
-        "@types/react-document-title": "^2.0.5",
-        "@types/react-dom": "^17.0.9",
-        "@types/react-router-dom": "^5.1.8",
-        "am-editable": "^0.0.6",
-        "animate.css": "^4.1.1",
-        "antd": "^4.16.13",
-        "antd-etable": "^1.3.71",
-        "antd-scss-theme-plugin": "^1.0.8",
-        "axios": "^0.24.0",
-        "babel-loader": "^8.2.2",
-        "clean-webpack-plugin": "^4.0.0",
-        "clipboard": "^2.0.10",
-        "copy-webpack-plugin": "^10.2.4",
-        "cross-env": "^7.0.3",
-        "css-loader": "^6.2.0",
-        "devtool": "^2.3.1",
-        "events": "^3.3.0",
-        "file-loader": "^6.2.0",
-        "html-webpack-plugin": "^5.3.2",
-        "less": "^4.1.1",
-        "less-loader": "^10.0.1",
-        "lodash": "^4.17.21",
-        "moment": "^2.29.1",
-        "plop": "^2.7.4",
-        "postcss": "^8.3.6",
-        "postcss-loader": "^6.1.1",
-        "prop-types": "^15.7.2",
-        "react": "^17.0.2",
-        "react-document-title": "^2.0.3",
-        "react-dom": "^17.0.2",
-        "react-hot-loader": "^4.13.0",
-        "react-pdf": "^5.6.0",
-        "react-redux": "^7.2.6",
-        "react-router-breadcrumbs-hoc": "^4.1.0",
-        "react-router-dom": "^5.3.0",
-        "react-thunk": "^1.0.0",
-        "react-transition-group": "^4.4.2",
-        "react-viewer": "^3.2.2",
-        "redux": "^4.1.2",
-        "redux-logger": "^3.0.6",
-        "redux-thunk": "^2.4.1",
-        "style-loader": "^3.2.1",
-        "typescript": "^4.4.2",
-        "webpack": "^5.52.0",
-        "webpack-cli": "^4.8.0",
-        "webpack-dev-server": "^4.4.0",
-        "webpack-merge": "^5.8.0",
-        "xss": "^1.0.11"
-    },
-    "devDependencies": {
-        "@hot-loader/react-dom": "^17.0.1+4.13.0",
-        "@types/react-pdf": "^5.0.9",
-        "@types/redux-logger": "^3.0.9",
-        "axios": "^0.24.0",
-        "jsencrypt": "^3.2.1",
-        "prettier": "2.3.2",
-        "sass": "^1.39.0",
-        "sass-loader": "^12.3.0",
-        "webpack": "^5.52.0"
-    }
+  "name": "react-pro",
+  "version": "1.0.0",
+  "description": "react-pro-app",
+  "author": "koringz",
+  "license": "ISC",
+  "scripts": {
+    "open": "yarn go --open",
+    "go": "webpack-dev-server --config config/webpack.development.config.js",
+    "build:dev": "cross-env FIRST_ENV=development --env.NODE_ENV=development webpack --config config/webpack.development.config.js",
+    "build:prod": "cross-env FIRST_ENV=production --env.NODE_ENV=production webpack --config config/webpack.production.config.js",
+    "build:test": "cross-env FIRST_ENV=testing --env.NODE_ENV=testing webpack --config config/webpack.testing.config.js",
+    "polp:build": "polp pages `call name generator pages`",
+    "upgrade": "yarn upgrade-interactive --latest"
+  },
+  "dependencies": {
+    "@babel/plugin-syntax-typescript": "^7.17.10",
+    "@babel/preset-env": "^7.17.10",
+    "@babel/preset-react": "^7.16.7",
+    "@babel/preset-typescript": "^7.16.7",
+    "@types/react": "^18.0.8",
+    "@types/react-document-title": "^2.0.5",
+    "@types/react-dom": "^18.0.3",
+    "@types/react-router-dom": "^5.3.3",
+    "animate.css": "^4.1.1",
+    "antd": "^4.20.2",
+    "antd-scss-theme-plugin": "^1.0.8",
+    "axios": "^0.27.2",
+    "babel-loader": "^8.2.5",
+    "clean-webpack-plugin": "^4.0.0",
+    "copy-webpack-plugin": "^10.2.4",
+    "cross-env": "^7.0.3",
+    "css-loader": "^6.7.1",
+    "file-loader": "^6.2.0",
+    "html-webpack-plugin": "^5.5.0",
+    "less": "^4.1.2",
+    "less-loader": "^10.2.0",
+    "plop": "^3.1.0",
+    "postcss": "^8.4.13",
+    "postcss-loader": "^6.2.1",
+    "react": "^18.1.0",
+    "react-document-title": "^2.0.3",
+    "react-dom": "^18.1.0",
+    "react-hot-loader": "^4.13.0",
+    "react-redux": "^8.0.1",
+    "react-router-breadcrumbs-hoc": "^4.1.0",
+    "react-router-dom": "^6.3.0",
+    "react-transition-group": "^4.4.2",
+    "style-loader": "^3.3.1",
+    "typescript": "^4.6.4",
+    "webpack": "^5.72.0",
+    "webpack-cli": "^4.9.2",
+    "webpack-dev-server": "^4.9.0",
+    "webpack-merge": "^5.8.0",
+    "xss": "^1.0.11"
+  },
+  "devDependencies": {
+    "@babel/core": "^7.17.10",
+    "@hot-loader/react-dom": "^17.0.2+4.13.0",
+    "copy-webpack-plugin": "^10.2.4",
+    "prettier": "^2.6.2",
+    "redux": "^4.2.0",
+    "redux-logger": "^3.0.6",
+    "redux-thunk": "^2.4.1",
+    "sass": "^1.51.0",
+    "sass-loader": "^12.6.0",
+    "webpack": "^5.72.0"
+  }
 }
